pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
# package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
slack-bolt = "^1.18.1"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
slack-sdk = "^3.27.1"
googletrans = "3.1.0a0"
langdetect = "^1.0.9"
google-api-python-client = "^2.129.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"